requirements.txt
flask
diffusers
torch
transformers
Pillow
python-dotenv
